feat: add more functionallity
