+++ src/UncontroledForm/UncontrolledForm.module.css
@@ -1,26 +1,37 @@
+.container {
+    width: 1000px;
+}
+
 .input {
- width: 300px;
+ width: 500px;
  height: 30px;
- margin-bottom: 20px;
+ border-radius: 10px;
+ border: 3px solid violet;
+ padding-left: 15px
+}
+
+.inputcountry {
+ width: 500px;
+ height: 30px;
+
  border-radius: 10px;
  border: 3px solid violet;
  padding-left: 15px
 }
 
 .inputsex {
-  width: 50px;
-  margin-bottom: 25px;
+  width: 20px;
+  margin-bottom: 15px;
 }
 
 .inputphoto {
-    border: none;
-    margin-top: 5px;
-    margin-bottom: 25px;
+ border: none;
+ width: 500px;
+ height: 30px;
 }
 
 .inputagree {
     width: 20px;
-    margin-bottom: 25px;
 }
 
 .inputsubmit {
@@ -32,7 +43,41 @@
     padding: 0;
 }
 
-/* .forerror {
-    font-size: 10px;
-    padding: -35px;
-}  */
+.inputwrapper {
+    display: flex;
+    flex-direction: column;
+}
+
+.inputwrapper1 {
+    display: flex;
+    flex-wrap: wrap;
+    width: 500px;
+    height: 35px;
+    margin-bottom: 20px;
+}
+
+.error {
+    width: 400px;
+    font-size: 14px;
+    color: green;
+    height: 25px;
+}
+
+.label {
+    margin-bottom: 10px;
+}
+
+.label1 {
+    font-size: 12px;
+    font-weight: bold;
+    color: rgb(33, 53, 71);
+}
+
+.disabled {
+    opacity: 0.5;
+    pointer-events: none
+}
+
+.submit {
+    pointer-events: auto
+}
